@@ -52,9 +52,9 @@
   },
   "homepage": "https://github.com/skjnldsv/nextcloud-upload",
   "devDependencies": {
-    "@cypress/vue2": "^2.1.0",
+    "@cypress/vue": "^6.0.0",
     "@nextcloud/eslint-config": "^8.3.0",
-    "@nextcloud/vite-config": "^1.2.2",
+    "@nextcloud/vite-config": "github:nextcloud-libraries/nextcloud-vite-config#vue3",
     "@tsconfig/cypress": "^1.0.2",
     "@types/node": "^20.11.30",
     "@vitest/coverage-v8": "^1.5.2",
@@ -72,7 +72,7 @@
   "dependencies": {
     "@nextcloud/auth": "^2.2.1",
     "@nextcloud/axios": "^2.4.0",
-    "@nextcloud/dialogs": "^5.2.0",
+    "@nextcloud/dialogs": "github:nextcloud-libraries/nextcloud-dialogs#vue3",
     "@nextcloud/files": "^3.1.1",
     "@nextcloud/l10n": "^3.0.1",
     "@nextcloud/logger": "^3.0.1",
@@ -86,7 +86,7 @@
     "simple-eta": "^3.0.2"
   },
   "peerDependencies": {
-    "@nextcloud/vue": "^8.0.0-beta || ^8.0.0",
-    "vue": "^2.7.16"
+    "@nextcloud/vue": "^9.0.0-alpha.2",
+    "vue": "^3.4.27"
   }
 }
