@@ -86,7 +86,7 @@ import type { PropType } from 'vue'
 import { getNewFileMenuEntries, Folder } from '@nextcloud/files'
 import { showError } from '@nextcloud/dialogs'
 import makeEta from 'simple-eta'
-import Vue from 'vue'
+import { defineComponent } from 'vue'
 
 import NcActionButton from '@nextcloud/vue/dist/Components/NcActionButton.js'
 import NcActions from '@nextcloud/vue/dist/Components/NcActions.js'
@@ -104,7 +104,7 @@ import { Status as UploadStatus } from '../upload.ts'
 import { t } from '../utils/l10n.ts'
 import logger from '../utils/logger.ts'
 
-export default Vue.extend({
+export default defineComponent({
 	name: 'UploadPicker',
 
 	components: {
